requirements-extra.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp[speedups]==3.9.3
aiohttp-jinja2==1.6
aiohttp-swagger3==0.8.0
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
apscheduler==3.10.4
asn1crypto==1.5.1
attrs==23.2.0
autocommand==2.2.2
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
backoff==2.2.1
bcrypt==4.1.2
boto3==1.34.44
botocore==1.34.44
brotli==1.1.0
cachetools==5.3.2
cattrs==23.2.3
cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.0
cherrypy==18.9.0
cryptography==41.0.7
deepdiff==6.7.1
defusedxml==0.7.1
deprecated==1.2.14
detect-secrets==1.4.0
fastjsonschema==2.19.1
filelock==3.13.1
frozendict==2.4.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
inflect==7.0.0
isodate==0.6.1
jaraco-collections==5.0.0
jaraco-context==4.3.0
jaraco-functools==4.0.0
jaraco-text==3.12.0
jinja2==3.1.3
jmespath==1.0.1
jq==1.6.0
jsons==1.6.3
kubernetes==29.0.0
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onelogin==2.0.4
ordered-set==4.1.0
packaging==23.2
paramiko==3.4.0
parsy==2.1
pint==0.23
plantuml==0.3.0
platformdirs==3.11.0
portalocker==2.8.2
portend==3.2.0
posthog==3.4.1
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic-core==2.16.2
pygithub==2.2.0
pygments==2.17.2
pyjwt[crypto]==2.8.0
pymysql==1.1.0
pynacl==1.5.0
pyopenssl==23.3.0
pyparsing==3.1.1
python-arango==7.9.1
python-dateutil==2.8.2
pytz==2024.1
pyvmomi==8.0.2.0.1
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
fixinventoryclient==2.0.0
fixinventorydata==0.2.4
fixdatalink[extra]==2.0.1
resotoui==3.7.1
retrying==1.3.4
rfc3339-validator==0.1.4
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
setuptools==69.1.0
six==1.16.0
slack-sdk==3.27.0
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
sqlalchemy==1.4.51
tempora==5.5.1
tenacity==8.2.3
tomlkit==0.12.3
toolz==0.12.1
transitions==0.9.0
typeguard==4.1.5
typing-extensions==4.9.0
typish==1.9.3
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
ustache==0.1.5
wcwidth==0.2.13
websocket-client==1.7.0
wrapt==1.16.0
yarl==1.9.4
zc-lockfile==3.0.post1
zipp==3.17.0
